refactor(edge/types): renamed and refactored types module to runtime module
